add the commands to support stepout, variables view, stack trace view
There is an additional change to handle setting the break point in a
file at a line number as two steps rather then one.  Needed to restore
break points.  It was always missing the first break point of a file as
it was opening it.  So, I separated the commands and it worked better.