add the commands to be sent to the debugger for step out, variables view and stack trace view
step out sends an 'r'
variables view sends a 'y'
stackTrace view sends a 'T'
Also separated out the file open and setting break line as a I needed
to have them separate when loading break points from a file.  It seemed
to miss the first break point of a file because it was just opening it.
 So, I separated the actions and it seemed to set them all correctly.