Add a mapping to start the debugger with the arguments under the cursor
Added a method that will take the line under the cursor and use it as
the arguments to the debugger and then start the debugger.  It will
ignore white space and a pound sign (ie, the comment character in Perl)
and take the rest of the line and pass it into the debugger.  This way
you can put the arguments in the top of the file you are debugging, and
press <Leader>d and it will start up the debugger with those arguments.
An example in the perl file you want to debug put something like
# -f  theFileName.txt -i 234 -d /home/foobar/version20
Then instead of entering the arguments each time just edit this line,
or have multlipe lines with the stuff you want an it will execute it
with those arguments.