add the ability to edit the break points file
Add the ability to edit the break points file separate from the
debugging session.  Using the <leader>be (breakpointsfile edit) will
open a editing buffer with the special break points file.  Then using
the <leader>o (defined only in this buffer) will split the window and
show the file under the cursor at the line number specified in the
breakpoint line.  If it is not in the correct place (ie, an executable
line) then ctl-w k (move window up to the break points edit buffer) and
change the number to something appropriate or comment out the line by
inserting a pound sign in the first column of the line.  Think of the
pound sign as disabling the breakpoint rather then removing the line
which would remove the breakpoint. This way you can keep the
breakpoints file update to date.